+++ Krowi_AchievementFilter/GUI/AchievementButton/AchievementButton.lua
@@ -5,28 +5,29 @@ local gui = addon.GUI;
 gui.AchievementButton = {};
 local achievementButton = gui.AchievementButton;
 
-local OnEnter, OnLeave, ShowTooltip;
-
-function achievementButton.PostLoadButtons(achievementsFrame)
+local OnEnter, OnLeave, OnClick;
+function achievementButton:PostLoadButtons(achievementsFrame)
+	-- Here we hook a lot of our own functionality to extend the default Blizzard Buttons
 	diagnostics.Trace("achievementButton.PostLoadButtons");
 
 	for _, button in next, achievementsFrame.Container.buttons do
 		button:RegisterForClicks("LeftButtonUp", "RightButtonUp");
 		button.Click = function(self, button, down, ignoreModifiers, anchor, offsetX, offsetY)
-			achievementButton.OnClick(self, button, achievementsFrame, ignoreModifiers, anchor, offsetX, offsetY);
+			OnClick(self, button, achievementsFrame, ignoreModifiers, anchor, offsetX, offsetY);
 		end;
 		button:SetScript("OnClick", button.Click);
-
-		button.shield:RegisterForClicks("LeftButtonUp", "RightButtonUp");
-		button.shield:SetScript("OnClick", function(self, button, down)
-			local parent = self:GetParent();
-			parent.Click(parent, button, achievementsFrame);
-		end);
 
 		if not achievementsFrame.UIFontHeight then
 			local _, fontHeight = button.description:GetFont();
 			achievementsFrame.UIFontHeight = fontHeight;
 		end
+
+		-- Change display behaviour
+		button.DisplayObjectives = function(button, renderOffScreen)
+			return self.Display.DisplayObjectives(button, renderOffScreen, achievementsFrame);
+		end;
+		button.ResetMetas = self.Display.ResetMetas;
+		button.ResetCriteria = self.Display.ResetCriteria;
 
 		-- Change tooltip behaviour
 		button:HookScript("OnEnter", function(self)
@@ -36,176 +37,18 @@ function achievementButton.PostLoadButtons(achievementsFrame)
 			OnLeave(self, achievementsFrame);
 		end);
 		button.shield:EnableMouse(false);
-		button.ShowTooltip = ShowTooltip;
+		button.ShowTooltip = function()
+			self.Tooltip.ShowTooltip(button);
+		end;
 	end
 end
 
 -- [[ OnEnter ]] --
-local AddBlizzardDefault, AddPartOfAChain, AddRequiredFor;
 function OnEnter(self, achievementsFrame)
 	diagnostics.Trace("OnEnter");
 
 	achievementsFrame.SetHighlightedButton(self);
-	ShowTooltip(self);
-end
-
-function ShowTooltip(button)
-	diagnostics.Trace("ShowTooltip");
-
-	GameTooltip:SetOwner(button, "ANCHOR_NONE");
-	GameTooltip:SetPoint("TOPLEFT", button, "TOPRIGHT");
-
-	AddBlizzardDefault(button);
-	AddPartOfAChain(button);
-	AddRequiredFor(button);
-
-	-- testing
-	-- local crits = GetAchievementNumCriteria(self.Achievement.ID);
-	-- for i = 1, crits do
-	-- 	diagnostics.DebugTable({GetAchievementCriteriaInfo(self.Achievement.ID, i)});
-	-- end
-	-- end testing
-
-	-- AchievementFrameAchievements_CheckGuildMembersTooltip(self.shield); -- Guild can be ignored for now
-
-	GameTooltip:Show();
-end
-
-local function AddAchievementLine(self, id)
-	local _, name, _, completed, _, _, _, _, _, _, _, _, wasEarnedByMe = addon.GetAchievementInfo(id);
-	local r, g, b;
-	if id == self.Achievement.ID then
-		if completed then
-			r, g, b = 0.75, 1, 0.75;
-		else
-			r, g, b = 0.9, 0.9, 0.9;
-		end
-	elseif completed then
-		r, g, b = 0.25, 0.75, 0.25;
-	else
-		r, g, b = 0.6, 0.6, 0.6;
-	end
-	GameTooltip:AddLine(name, r, g, b); -- Achievement name
-	local tooltipTextureInfo;
-	if addon.Options.db.Tooltip.ShowCurrentCharacterIcons then
-		if wasEarnedByMe then
-			GameTooltip:AddTexture(136814); -- Check texture
-		elseif self.Achievement.NotObtainable then
-			GameTooltip:AddTexture(136813); -- Fail texture
-		else
-			GameTooltip:AddTexture(136815); -- Dash texture
-		end
-		tooltipTextureInfo = {margin = { right = 24}};
-	end
-	if completed then
-		GameTooltip:AddTexture(136814, tooltipTextureInfo); -- Check texture
-	elseif self.Achievement.NotObtainable then
-		GameTooltip:AddTexture(136813, tooltipTextureInfo); -- Fail texture
-	else
-		GameTooltip:AddTexture(136815, tooltipTextureInfo); -- Dash texture
-	end
-end
-
-function AddBlizzardDefault(self)
-	diagnostics.Trace("AddBlizzardDefault");
-
-	if ( self.accountWide ) then
-		if ( self.completed ) then
-			GameTooltip:AddLine(ACCOUNT_WIDE_ACHIEVEMENT_COMPLETED);
-		else
-			GameTooltip:AddLine(ACCOUNT_WIDE_ACHIEVEMENT);
-		end
-		GameTooltip:Show();
-		return;
-	end
-	if ( self.shield.earnedBy ) then
-		GameTooltip:AddLine(format(ACHIEVEMENT_EARNED_BY,self.shield.earnedBy));
-		local me = UnitName("player")
-		if ( not self.shield.wasEarnedByMe ) then
-			GameTooltip:AddLine(format(ACHIEVEMENT_NOT_COMPLETED_BY, me));
-		elseif ( me ~= self.shield.earnedBy ) then
-			GameTooltip:AddLine(format(ACHIEVEMENT_COMPLETED_BY, me));
-		end
-		GameTooltip:Show();
-		return;
-	end
-end
-
-function AddPartOfAChain(self)
-	if not self.Achievement or not self.Achievement.ID or (not GetNextAchievement(self.Achievement.ID) and not GetPreviousAchievement(self.Achievement.ID)) or not addon.Options.db.Tooltip.ShowPartOfAChain then
-		return;
-	end
-
-	local id;
-	local tmpID = self.Achievement.ID;
-	while tmpID do -- Find first achievement in a chain
-		id = tmpID;
-		tmpID = GetPreviousAchievement(id);
-	end
-
-	if (GameTooltip:NumLines() > 0) then
-		GameTooltip:AddLine(" "); -- Empty line to seperate it from the previous block
-	end
-	GameTooltip:AddLine(addon.L["Part of a chain"]); -- Header
-
-	self.Achievement:ClearPartOfAChainIDs(); -- Make sure it's empty before adding new stuff
-	while id do
-		AddAchievementLine(self, id);
-		self.Achievement:AddPartOfAChainID(id);
-		id = GetNextAchievement(id);
-	end
-end
-
-local criteriaCache = {};
-local criteriaCacheIsEmpty = true;
-function AddRequiredFor(self)
-	if not addon.Options.db.Tooltip.ShowRequiredFor then
-		return;
-	end
-
-	if criteriaCacheIsEmpty then -- Build cache the first time
-		local nils, i = 0, 1;
-		while nils < 500 do -- Biggest gap is 209 in 9.0.5 as of 2021-05-03
-			local id = addon.GetAchievementInfo(i);
-			if id then
-				local numCriteria = GetAchievementNumCriteria(id);
-				if numCriteria > 0 then
-					for j = 1, numCriteria do
-						local _, criteriaType, _, _, _, _, _, assetID = GetAchievementCriteriaInfo(id, j);
-						if criteriaType == 8 then
-							tinsert(criteriaCache, {AchievementID = assetID, RequiredForID = id});
-						end
-					end
-				end
-				nils = 0;
-			else
-				nils = nils + 1;
-			end
-			i = i + 1;
-		end
-		criteriaCacheIsEmpty = nil;
-	end
-
-	local ids = {};
-	for _, criteria in next, criteriaCache do
-		if criteria.AchievementID == self.Achievement.ID then
-			-- local id = addon.GetAchievementInfo(criteria.RequiredForID);
-			tinsert(ids, criteria.RequiredForID);
-		end
-	end
-
-	if #ids > 0 then
-		if (GameTooltip:NumLines() > 0) then
-			GameTooltip:AddLine(" "); -- Empty line to seperate it from the previous block
-		end
-		GameTooltip:AddLine(addon.L["Required for"]); -- Header
-
-		self.Achievement:ClearRequiredForIDs(); -- Make sure it's empty before adding new stuff
-		for _, id in next, ids do
-			AddAchievementLine(self, id);
-			self.Achievement:AddRequiredForID(id);
-		end
-	end
+	self.ShowTooltip();
 end
 
 -- [[ OnLeave ]] --
@@ -219,7 +62,7 @@ end
 
 -- [[ OnClick ]] --
 local OnClickLeftButton, OnClickRightButton;
-function achievementButton.OnClick(self, button, achievementsFrame, ignoreModifiers, anchor, offsetX, offsetY) -- ignoreModifiers, anchor, offsetX, offsetY are used for in code calls
+function OnClick(self, button, achievementsFrame, ignoreModifiers, anchor, offsetX, offsetY) -- ignoreModifiers, anchor, offsetX, offsetY are used for in code calls
 	diagnostics.Trace("achievementButton.OnClick");
 
 	if button == "LeftButton" then