+++ Krowi_AchievementFilter/GUI/Search/FullSearchResultsButton.lua
@@ -1,21 +1,23 @@
 -- [[ Namespaces ]] --
 local _, addon = ...;
 local diagnostics = addon.Diagnostics;
-local gui = addon.GUI;
-gui.FullSearchResult = {};
-local fullSearchResult = gui.FullSearchResult;
+local search = addon.GUI.Search;
+search.FullSearchResult = {};
+local fullSearchResult = search.FullSearchResult;
 
+local OnClick;
 function fullSearchResult.PostLoadButtons(frame, achievementsFrame)
-	diagnostics.Trace("fullSearchResult.PostLoadButtons");
+	-- Here we hook a lot of our own functionality to extend the default Blizzard Buttons
+	diagnostics.Trace("fullSearchResult:PostLoadButtons");
 
     for _, button in next, frame.Container.buttons do
 		button:SetScript("OnClick", function(self, button, down)
-			gui.FullSearchResult.OnClick(self, frame, achievementsFrame);
+			OnClick(self, frame, achievementsFrame);
 		end);
 	end
 end
 
-function fullSearchResult.OnClick(self, fullSearchResultsFrame, achievementsFrame)
+function OnClick(self, fullSearchResultsFrame, achievementsFrame)
     diagnostics.Trace("fullSearchResult.OnClick");
 
     if self.Achievement then